add AX_HARDEN_CC_COMPILER_FLAGS and warning fixes

+++ src/certgen/clu_certgen_ecc.c
@@ -51,7 +51,7 @@ int make_self_signed_ecc_certificate(char* keyPath, char* certOut, int oid) {
     }
 
     fseek(keyFile, 0, SEEK_END);
-    keyFileSz = ftell(keyFile);
+    keyFileSz = (int)ftell(keyFile);
     byte keyBuf[keyFileSz];
     fseek(keyFile, 0, SEEK_SET);
     fread(keyBuf, 1, keyFileSz, keyFile);
